TabController: tab controller added to manage overview/Exercises/Muscles screens
