+++ lib/screens/chest_screen.dart
@@ -1,13 +1,11 @@
+import 'package:bubble_tab_indicator/bubble_tab_indicator.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter/painting.dart';
-import 'package:training/components/activity_card.dart';
 import 'package:training/components/app_bar_button.dart';
-import 'package:training/components/chest_menu_item.dart';
-import 'package:training/components/circular_icon.dart';
-import 'package:training/components/health_radio.dart';
-import 'package:training/components/number_card.dart';
-import 'package:training/components/simple_pie_chart.dart';
 import 'package:training/constants.dart';
+import 'package:training/screens/exercises_screen.dart';
+import 'package:training/screens/muscles_screen.dart';
+import 'package:training/screens/overview_screen.dart';
 import 'package:training/screens/right_menu_screen.dart';
 
 class ChestScreen extends StatelessWidget {
@@ -82,146 +80,41 @@ class ChestScreen extends StatelessWidget {
               ),
             ),
             Expanded(
-              child: Row(
-                mainAxisAlignment: MainAxisAlignment.spaceEvenly,
-                children: [
-                  ChestMenuItem(
-                    text: 'Overview',
-                    textColor: kTextChestMenuActiveColor,
-                    bgColor: kBtnChestMenuActiveColor,
-                  ),
-                  ChestMenuItem(
-                    text: 'Exercises',
-                    textColor: kTextChestMenuInActiveColor,
-                    bgColor: kBtnChestMenuInActiveColor,
-                  ),
-                  ChestMenuItem(
-                    text: 'Muscles',
-                    textColor: kTextChestMenuInActiveColor,
-                    bgColor: kBtnChestMenuInActiveColor,
-                  ),
-                ],
-              ),
-            ),
-            Expanded(
-              flex: 2,
-              child: Row(
-                children: [
-                  Expanded(
-                    flex: 3,
-                    child: Container(
-                      width: 200,
-                      height: 200,
-                      child: SimplePieChart.withSampleData(),
-                    ),
-                  ),
-                  Expanded(
-                    flex: 2,
-                    child: Column(
-                      mainAxisAlignment: MainAxisAlignment.spaceEvenly,
-                      children: [
-                        HealthRadio(
-                          label: 'Cardio',
-                          iconColor: Color(0xFFFF5722),
-                        ),
-                        HealthRadio(
-                          label: 'Strength',
-                          iconColor: Color(0xFF6E22E5),
-                        ),
-                        HealthRadio(
-                          label: 'Stretch',
-                          iconColor: Color(0xFF22A1E5),
-                        ),
+              flex: 5,
+              child: DefaultTabController(
+                length: 3,
+                child: Scaffold(
+                  appBar: AppBar(
+                    elevation: 0.0,
+                    toolbarHeight: 50.0,
+                    bottom: TabBar(
+                      indicatorSize: TabBarIndicatorSize.tab,
+                      labelColor: kTextChestMenuActiveColor,
+                      unselectedLabelColor: kTextChestMenuInActiveColor,
+                      indicator: BubbleTabIndicator(
+                        indicatorHeight: 50.0,
+                        indicatorColor: Color(0xFFF4F4F7),
+                        tabBarIndicatorSize: TabBarIndicatorSize.tab,
+                        indicatorRadius: 10,
+                        insets: EdgeInsets.all(0),
+                      ),
+                      tabs: [
+                        Text('Overview'),
+                        Text('Exercises'),
+                        Text('Muscles'),
                       ],
                     ),
                   ),
-                  Expanded(
-                    child: Column(
-                      mainAxisAlignment: MainAxisAlignment.spaceEvenly,
-                      children: [
-                        NumberCard(label: '18', unit: '%'),
-                        NumberCard(label: '53', unit: '%'),
-                        NumberCard(label: '29', unit: '%'),
-                      ],
-                    ),
-                  )
-                ],
-              ),
-            ),
-            Expanded(
-              child: Row(
-                mainAxisAlignment: MainAxisAlignment.spaceEvenly,
-                children: [
-                  CircularIcon(
-                    icon: Icons.bolt,
-                    iconColor: Color(0xFFFF5722),
-                    bgColor: Color(0x15FF5722),
-                  ),
-                  CircularIcon(
-                    icon: Icons.sports_basketball_sharp,
-                    iconColor: Color(0xFF6E22E5),
-                    bgColor: Color(0x156E22E5),
-                  ),
-                  CircularIcon(
-                    icon: Icons.timer,
-                    iconColor: Color(0xFF22A1E5),
-                    bgColor: Color(0x1522A1E5),
-                  ),
-                ],
-              ),
-            ),
-            Expanded(
-              child: Row(
-                mainAxisAlignment: MainAxisAlignment.spaceEvenly,
-                children: [
-                  Column(
-                    mainAxisAlignment: MainAxisAlignment.spaceEvenly,
+                  body: TabBarView(
                     children: [
-                      NumberCard(label: '450', unit: 'cal'),
-                      ActivityCard(label: 'Burned'),
+                      OverviewScreen(),
+                      ExercisesScreen(),
+                      MusclesScreen(),
                     ],
                   ),
-                  Column(
-                    mainAxisAlignment: MainAxisAlignment.spaceEvenly,
-                    children: [
-                      NumberCard(label: '148', unit: 'kg'),
-                      ActivityCard(label: 'Lifted'),
-                    ],
-                  ),
-                  Column(
-                    mainAxisAlignment: MainAxisAlignment.spaceEvenly,
-                    children: [
-                      NumberCard(label: '50', unit: 'min'),
-                      ActivityCard(label: 'Duration'),
-                    ],
-                  ),
-                ],
+                ),
               ),
             ),
-            Container(
-              decoration: BoxDecoration(
-                color: Color(0xFFFF5722),
-                borderRadius: BorderRadius.circular(20.0),
-              ),
-              padding: EdgeInsets.all(30.0),
-              child: Row(
-                mainAxisAlignment: MainAxisAlignment.spaceEvenly,
-                children: [
-                  Text(
-                    'Repeat Workout',
-                    style: TextStyle(
-                        fontWeight: FontWeight.w500,
-                        fontSize: 20.0,
-                        color: Colors.white),
-                  ),
-                  Icon(
-                    Icons.arrow_forward,
-                    size: 20.0,
-                    color: Colors.white,
-                  )
-                ],
-              ),
-            )
           ],
         ),
       ),
