+++ lib/screens/muscles_screen.dart
@@ -1,0 +1,8 @@
+import 'package:flutter/material.dart';
+
+class MusclesScreen extends StatelessWidget {
+  @override
+  Widget build(BuildContext context) {
+    return Icon(Icons.directions_car);
+  }
+}
