+++ pubspec.yaml
@@ -15,6 +15,7 @@ dependencies:
   cupertino_icons: ^1.0.2
   font_awesome_flutter: ^9.0.0
   charts_flutter: ^0.10.0
+  bubble_tab_indicator: ^0.1.6
 
 dev_dependencies:
   flutter_test:
